// h5/src/apis/models/ProjectPreEvaluationVo.ts
export interface ProjectPreEvaluationVo {
    /**
     * 项目ID
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    id?: string;
    /**
     * 项目名称
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    projectName?: string;
    /**
     * 所属板块（园区名称）
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    parkName?: string;
    /**
     * 国民经济分类
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    industryClassification?: string;
    /**
     * 投资方
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    investor?: string;
    /**
     * 投资总额（亿元）
     * @type {number}
     * @memberof ProjectPreEvaluationVo
     */
    totalInvestmentCny?: number;
    /**
     * 园区ID
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    park?: string;
    /**
     * 评估状态：未完成/已完成
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    evaluationStatus?: string;
    /**
     * 年度（查询条件）
     * @type {number}
     * @memberof ProjectPreEvaluationVo
     */
    year?: number;
    /**
     * 产业类型（查询条件）
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    projectType?: string;
    /**
     * 部门名称（查询条件）
     * @type {string}
     * @memberof ProjectPreEvaluationVo
     */
    deptName?: string;
}